Add TokenKind and Token to lexer.
I also added a Display impl for TokenKind.

Rust has a concept of "trait"s, which you can
kind of think of like interfaces that we can
tack onto existing structs.

The "Display" trait allows you to print the
struct, and provides a `.to_string()` method.

This is helpful for TokenKinds -- when we lex
into Tokens we may do something like:

```rs
if (/* matches left curly bracket, for instance */) {
    let kind = TokenKind::LeftCurly;

    return Token {
        kind: kind,
        lexeme: kind.to_string(),
    }
}
```

Note that for STRINGs, NUMBERs, and IDENTIFIERs, we can't
know the string representation of their kind ahead of time,
so we are using "STRING", "NUMBER", and "IDENTIFIER"
respectively. For all other tokens, though, this works as a
shortcut.