Add more unit tests.
As much as I'm generally iffy on writing
unit tests this early, writing a language
is an excellent case for test-driven-development
because of the regularity of the input; we
pretty much know ahead of time what the
lexer is going to be able to lex, what the
parser is going to be able to parse, etc.

I'm using a macro in the lexer test suite to make
tests easier to write. It takes a Token iterator
and checks that all the tokens match the lexemes
we provide. Macros might look a little scary at
first but they're really not that bad, and are
quite powerful.