Adds lexer functionality
This commit adds full functionality for the lexer. Now, a string can be
lexed using Lexer::lex() to generate a Vec<Token>. Thus far, it passes
all the tests however we may need to add more to ensure we cover all
edge cases etc. as currently there are only 8 tests.